devfreq_boost: add lower devfreq boost freq
We don't want to boost devfreq all the way af we don't
have heavy CPU load. Add variable and initially set it
to 3143 (595200 mhz for DDR bus), which is significantly
lower then standart 4173 (1324800 mhz). This results in
lower amount of heat and better battery life without
significant sacrifice to fluidity.

Signed-off-by: Yaroslav Furman <yaro330@gmail.com>